Separate CloudSync background runtime
Move retry, hook, result-accounting, and background synchronization execution behind the CloudSync lifecycle facade without changing runtime behavior.
